extension/backend/requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0

# Environment & Configuration
python-dotenv==1.0.0

# Database - Cloud SQL
psycopg2-binary==2.9.9
cloud-sql-python-connector[pg8000]==1.11.0

# Google Cloud Platform Services
google-cloud-aiplatform>=1.70.0
google-cloud-vision==3.5.0
google-cloud-speech==2.23.0
google-cloud-translate==3.13.0

# Google Authentication
google-auth>=2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google APIs
google-api-python-client==2.110.0

# Utilities
requests==2.31.0
protobuf==4.25.1